Rename Maths Project.R to Maths Dissertation Project.R
This was my first project carried out in RStudio. It was a statistical analysis surrounding linear regression, aided by Data visualisation. The topic was to create a linear mapping between housing prices in King County, Texas and many other variables such as number of bedrooms, bathrooms and other house features.